readme: Add a note about etcdfs
An etcdfs afero backend is available as part of the mgmt project at:
https://github.com/purpleidea/mgmt/tree/master/etcd/fs

I'll be maintaining it upstream there for now, but in the future we can
discuss moving it here if that makes sense. This note will hopefully aid
to tell interested users about the backend so that they don't have to
duplicate work if that's something they're interested in!