Allow BasePathFs to work with "." as a base path.
Currently this fails because RealPath checks that a path is local to the basepath by seeing if it's cleaned prefix starts with the basepath.

This doesn't work when the basepath is "." because Clean removes the "." prefix from the full path.

Instead we use a more robust check using Rel and IsLocal.

Also we remove the superflous call to Clean since Join already calls Clean on the result.